requirements.txt
# Core
streamlit>=1.29.0
langchain>=0.1.0

# Audio
SpeechRecognition>=3.10.0
pydub>=0.25.1
gTTS>=2.4.0
streamlit-audiorecorder>=0.0.6

# Vision
pillow>=10.1.0
opencv-python>=4.8.1
pytesseract>=0.3.10
colorthief>=0.2.1

# Utilities
plotly>=5.18.0
pandas>=2.1.4
numpy>=1.24.3
python-multipart>=0.0.6
streamlit-autorefresh>=0.4.0

# RAG
chromadb>=0.4.22
sentence-transformers>=2.2.2
pypdf>=3.17.0
python-docx>=1.1.0
